// src/app/enums/navigation-order.ts
export const enum NavigationOrder {
  DIRECTIVE_OPEN_EXTERNAL_LINKS = 1,
  DIRECTIVE_PASSWORD_STRENGTH = 2,
  NAIVE_CALENDAR = 3,
  DIRECTIVE_CALENDAR = 4,
  NAIVE_SELECT = 5,
  NAIVE_SELECT_AS_COMPONENT = 6,
  DIRECTIVE_SELECT = 7,
  NAIVE_USER_PERMISSION = 8,
  DIRECTIVE_USER_PERMISSION = 9,
  NAIVE_DISABLE_INTERACTIVE_ELEMENTS = 10,
  DIRECTIVE_DISABLE_INTERACTIVE_ELEMENTS = 11,
  NAIVE_CONTEXT_MENU = 12,
  DIRECTIVE_CONTEXT_MENU = 13,
  DIRECTIVE_COMPOSITION_API_CONTEXT_MENU = 100,
  DIRECTIVE_NG_PLURAL = 200,
}